Move getT into the shared fixture as well
